\name{iASeq internal}
\alias{f0.loglike}
\alias{fup.loglike}
\alias{fdown.loglike}
\alias{iASeqmotiffit}
\title{iASeq Internal Functions}
\description{These functions are not part of the package application programming interface and are not recommended to be used by the users.}

\usage{
f0.loglike
fup.loglike
fdown.loglike
iASeqmotiffit
}

\references{Yingying Wei, Xia Li, Qianfei Wang, Hongkai Ji (2012) iASeq: integrating multiple ChIP-seq datasets for detecting allele-specific binding.}